feat: set some functions to utils folder
