+++ react/graphql/uploadBrandTranslation.gql
